# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Memeloop Token Center API
  version: 0.1.0
  summary: Multi-protocol AI gateway, credential, quota, and archive API
  description: |
    Contract for the API surface implemented under `src/api/`.

    The production Helm deployment separates the **control** and **gateway** roles.
    `/internal/v1/*` is served only by the control role. `/self/v1/*` and `/v1/*`
    are served only by the gateway role. `/healthz` is present on every HTTP-serving
    role. `/livez` is process liveness and `/readyz` runs bounded PostgreSQL and
    archive probes. PostgreSQL determines the readiness HTTP status; archive
    failure is reported as degraded without withdrawing gateway endpoints.
    `/healthz` is a deprecated compatibility alias for `/livez`.

    All monetary amounts are decimal strings with at most six fractional digits.
    All timestamps named `*_at` are Unix epoch milliseconds unless stated otherwise.
    Successful and error responses under `/internal`, `/self`, and `/v1` include
    `X-MTC-API-Version: v1`.
    Issued client and service credential secrets are returned only by create/rotate
    operations and must be stored by the caller immediately.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: https://token-center.api.onetwo.website
    description: Gateway role
  - url: https://token-center.k3s.onetwo.website
    description: Control role
...
paths:
...
  /v1/responses:
    get:
      tags: [OpenAI compatible]
      operationId: negotiateResponsesWebSocket
      summary: Negotiate the Responses transport before native WebSocket support is available
      description: |
        Temporary authenticated compatibility response for clients that probe the
        Responses WebSocket endpoint before using HTTP. This operation returns 426
        immediately so the client can use `POST /v1/responses`. It does not remove or
        replace the planned native Responses WebSocket transport.
      x-transport-negotiation:
        lifecycle: temporary-until-native-websocket
        fallback-operation: POST /v1/responses
        native-websocket-planned: true
      security: [{ clientBearer: [] }]
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '426': { $ref: '#/components/responses/ResponsesWebSocketUpgradeRequired' }
    post:
      tags: [OpenAI compatible]
      operationId: createResponse
      summary: Proxy an OpenAI-compatible Responses request
      security: [{ clientBearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OpenAIResponseRequest' }
      responses:
        '200': { $ref: '#/components/responses/ProxySuccess' }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '424': { $ref: '#/components/responses/UnpricedModel' }
        '429': { $ref: '#/components/responses/UsageRejected' }
        '502': { $ref: '#/components/responses/UpstreamError' }
...
  /v1/images/generations:
    post:
      tags: [Multimodal generation]
      operationId: createImageGeneration
      summary: Generate an image synchronously or queue a provider-native job
      description: |
        When the body contains `input`, it uses the asynchronous generation contract and
        accepts optional `Idempotency-Key`: the first submission returns 202, an identical
        replay returns the same job with 200 without a second reservation, and reuse with
        different `model` or `input` returns 400. Without `input`, this endpoint proxies
        the synchronous OpenAI Images API or a configured Codex Responses image tool.

        Synchronous idempotency is isolated by the authenticated stable `key_id` and uses
        the canonical post-policy model/request payload. Authentication and traffic-policy
        evaluation precede the replay lookup, while route/price resolution and outbound
        client validation run only for a newly claimed execution. A completed same-payload
        replay therefore remains available after its route or price changes and returns the
        exact archived status, sanitized JSON body, and original request ID; it creates no
        request record, makes no upstream call, and consumes no additional quota/billing
        reservation, per-key rate-limit window, or image execution permit. Reusing the key
        with a different payload returns 400. An identical active request returns 409; an
        expired pending lease is atomically reclaimed for a fresh execution. For URL-mode
        results, Token Center downloads the object before success and persists, returns,
        and replays only a relative
        `/self/v1/requests/{request_id}/assets/{asset_id}` reference plus sanitized asset
        metadata—never the provider URL or internal archive locator.
      security: [{ clientBearer: [] }]
      x-idempotency-contract:
        namespace: stable-key-id
        fingerprint: post-policy-canonical-model-and-request
        completed-replay: exact-status-body-and-request-id
        different-payload-status: 400
        in-progress-status: 409
        replay-preconditions:
          - authentication
          - traffic-policy-evaluation
        fresh-execution-only:
          - route-resolution
          - price-lookup
          - outbound-client-validation
        expired-pending: atomically-reclaimed
        completed-replay-effects:
          - no-request-record
          - no-upstream-request
          - no-quota-reservation
          - no-usage-settlement
          - no-key-rate-limit-window-consumption
          - no-image-execution-permit
        persisted-url-results: mtc-request-asset-reference-only
      parameters: [{ $ref: '#/components/parameters/OptionalIdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/AsyncGenerationRequest'
                - $ref: '#/components/schemas/OpenAIImageRequest'
      responses:
        '200': { $ref: '#/components/responses/ImageSuccessOrGenerationReplay' }
        '202': { $ref: '#/components/responses/GenerationAccepted' }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '424': { $ref: '#/components/responses/UnpricedModel' }
        '429': { $ref: '#/components/responses/UsageRejected' }
        '502': { $ref: '#/components/responses/UpstreamError' }
components:
  securitySchemes:
    serviceBearer:
      type: http
      scheme: bearer
      bearerFormat: mts_…
      description: Bootstrap service token or a stored service credential with the required scope and tenant boundary.
    clientBearer:
      type: http
      scheme: bearer
      bearerFormat: mtc_… or imported opaque credential
      description: The active normal credential generation for one stable client identity.
    memeloopCloudHmac:
      type: apiKey
      in: header
      name: X-MTC-Webhook-Signature
      description: RustCrypto HMAC-SHA-256 signature over `timestamp + "." + exact_body`; requires the timestamp header documented on the operation.
...
  parameters:
...
    OptionalIdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string, minLength: 1, maxLength: 200, pattern: '^[!-~]+$' }
...
  responses:
...
    InvalidRequest:
      description: Invalid JSON, field, state transition, idempotency reuse, or unsupported operation.
      headers:
        X-MTC-Request-Id: { $ref: '#/components/headers/RequestId' }
        X-MTC-API-Version: { $ref: '#/components/headers/ApiVersion' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    ResponsesWebSocketUpgradeRequired:
      description: Native Responses WebSocket transport is not available in this release; use the authenticated HTTP POST operation.
      headers:
        Upgrade:
          description: Requested protocol token; this release responds with 426 instead of switching transports.
          schema: { type: string, const: websocket }
        X-MTC-Request-Id: { $ref: '#/components/headers/RequestId' }
        X-MTC-API-Version: { $ref: '#/components/headers/ApiVersion' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Missing, malformed, revoked, expired, or invalid bearer credential.
      headers:
        X-MTC-Request-Id: { $ref: '#/components/headers/RequestId' }
        X-MTC-API-Version: { $ref: '#/components/headers/ApiVersion' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Credential lacks the required scope, tenant boundary, or model permission.
      headers:
        X-MTC-Request-Id: { $ref: '#/components/headers/RequestId' }
        X-MTC-API-Version: { $ref: '#/components/headers/ApiVersion' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    UnpricedModel:
      description: Model has no price in the client's currency.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    UsageRejected:
      description: Rate limit, concurrency, balance, or configured budget rejected the request before upstream execution.
      headers:
        Retry-After:
          description: Seconds until the next meaningful retry for retryable dimensions. Permanent balance/lifetime failures omit this header.
          schema: { type: integer, minimum: 1 }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/UsageLimitErrorResponse' }
    UpstreamError:
      description: Configured upstream or OAuth provider was unavailable. Sensitive upstream detail is not returned.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    ProxySuccess:
      description: Upstream response. Content type and streaming semantics follow the selected protocol/provider.
      content:
        application/json:
          schema: { type: object, additionalProperties: true }
        text/event-stream:
          schema: { type: string }
    GenerationAccepted:
      description: Job was priced, reserved, archived, and queued.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/GenerationJob' }
...
    ImageSuccessOrGenerationReplay:
      description: |
        Sanitized synchronous image response, exact completed synchronous replay, or an
        existing asynchronous job returned by an idempotent replay. Synchronous URL-mode
        results contain only Token Center request-asset references, never provider URLs or
        internal archive locators.
      headers:
        X-MTC-Request-Id: { $ref: '#/components/headers/RequestId' }
        X-MTC-API-Version: { $ref: '#/components/headers/ApiVersion' }
      content:
        application/json:
          schema: { type: object, additionalProperties: true }
  schemas:
...
    OpenAIResponseRequest:
      type: object
      required: [model, input]
      additionalProperties: true
      properties:
        model: { type: string, minLength: 1 }
        input: true
        stream: { type: boolean }
        previous_response_id:
          type: [string, 'null']
          maxLength: 256
          description: |
            OpenAI Responses continuation identifier. Schema v16 stores a successful
            upstream response `id` on the prior logical-conversation observation, so a
            later request can use this value as direct parent evidence even when the
            client does not send Token Center-specific turn headers.
...
    AsyncGenerationRequest:
      type: object
      additionalProperties: false
      required: [model, input]
      properties:
        model: { type: string, minLength: 1, maxLength: 200 }
        input: { type: object, additionalProperties: true, description: Provider-native Seedance input, closed SiliconFlow T2V parameters, or administrator-owned ComfyUI workflow inputs. Consult the selected model's generation_schema when present. }
    OpenAIImageRequest:
      type: object
      required: [model, prompt]
      additionalProperties: true
      properties:
        model: { type: string, minLength: 1, maxLength: 200 }
        prompt: { type: string, minLength: 1, maxLength: 32000 }